Project List Window
Projects retrieved from the database are listed as interactive boxes. There is a button on the projects page that allows us to add a new project. The location of this button is updated so that it is at the end of the projects each time a project is added. Clicking on any project opens another window displaying detailed information about that project. This window contains two buttons for deleting or updating the project. When deleting, we'll be presented with a confirmation box. If we confirm, the project will be deleted and we'll return to the project listing page. If we don't, we'll remain on the project page. The update process hasn't been added yet.